+++ Source/gui/TitleButton.h
@@ -1,0 +1,32 @@
+/*
+** Odin 2 Synthesizer Plugin
+** Copyright (C) 2020 - 2021 TheWaveWarden
+**
+** Odin 2 is free software: you can redistribute it and/or modify
+** it under the terms of the GNU General Public License as published by
+** the Free Software Foundation, either version 3 of the License, or
+** (at your option) any later version.
+**
+** Odin 2 is distributed in the hope that it will be useful,
+** but WITHOUT ANY WARRANTY; without even the implied warranty of
+** MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+** GNU General Public License for more details.
+*/
+
+#pragma once
+
+#include "../JuceLibraryCode/JuceHeader.h"
+#include "../GlobalIncludes.h"
+
+#define TWW_HOMEPAGE_URL "https://thewavewarden.com"
+
+class TitleButton : public juce::Button {
+public:
+  TitleButton();
+
+  void paintButton(Graphics &g, bool shouldDrawButtonAsHighlighted, bool shouldDrawButtonAsDown) override;
+  void setGUIBig();
+  void setGUISmall();
+
+  juce::Image m_hover_image;
+};